Adding popups

+++ client/src/pages/Signin.tsx
@@ -49,24 +49,33 @@ export default function SignIn() {
   const [modal, setModal] = useState("");
   const [open, setOpen] = useState(false);
   const handleClose = () => setOpen(false);
-  
   const onSubmit: SubmitHandler<FormFeilds> = async (data) => {
-  try {
-    const result = await postRequest("/signin", data);
-    if (result.message !== "Invalid username or password") {
-      localStorage.setItem("denaurlen-token", JSON.stringify(result.token));
-      navigate("/friends");
-    } else {
+    try {
+      const result = await postRequest("/signin", data);
+      if (result.message !== "Invalid username or password") {
+        const existingToken = localStorage.getItem("denaurlen-token");
+        if (existingToken) {
+          window.alert("Token already exists in local storage");
+          return;
+        }
+        localStorage.setItem("denaurlen-token", JSON.stringify(result.token));
+
+        setOpen(true); // Open the modal
+        setModal("Login successful");
+        setTimeout(() => {
+          navigate("/friends");
+        }, 1000);
+      } else {
+        setOpen(true);
+        setModal(result.message);
+      }
+    } catch (error) {
+      console.error(error);
       setOpen(true);
-      setModal(result.message);
+      setModal("Something went wrong");
     }
-  } catch (error) {
-    console.error(error);
-    setOpen(true);
-    setModal("Something went wrong");
-  }
-  console.log(data);
-};
+    console.log(data);
+  };
   return (
     <Box
       sx={{
